requirements.txt
asgiref==3.4.1
boto3==1.18.32
botocore==1.21.32
Django==4.0
django-cleanup==5.2.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.13.1
factory-boy==3.2.0
Faker==11.3.0
jmespath==0.10.0
Pillow==9.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
gunicorn==20.1.0
djangorestframework-simplejwt==5.0.0
django-modeltranslation==0.17.5
drf-yasg==1.21.3